Rename and generalize rating box component(s)
User for user power and user influence.
